Move test support bp func to test package.
