+++ blockproducer/rpc.go
@@ -19,13 +19,10 @@ package blockproducer
 import (
 	"context"
 	"database/sql"
-	"fmt"
 	"strings"
 	"time"
 
 	pi "github.com/CovenantSQL/CovenantSQL/blockproducer/interfaces"
-	"github.com/CovenantSQL/CovenantSQL/crypto"
-	"github.com/CovenantSQL/CovenantSQL/crypto/asymmetric"
 	"github.com/CovenantSQL/CovenantSQL/proto"
 	"github.com/CovenantSQL/CovenantSQL/route"
 	"github.com/CovenantSQL/CovenantSQL/rpc"
@@ -184,80 +181,4 @@ func WaitDatabaseCreation(
 			return
 		}
 	}
-}
-
-// WaitBPChainService waits until BP chain service is ready.
-func WaitBPChainService(ctx context.Context, period time.Duration) (err error) {
-	var (
-		ticker = time.NewTicker(period)
-		req    = &types.FetchBlockReq{
-			Height: 0, // Genesis block
-		}
-	)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ticker.C:
-			if err = rpc.RequestBP(
-				route.MCCFetchBlock.String(), req, nil,
-			); err == nil || !strings.Contains(err.Error(), "can't find service") {
-				return
-			}
-		case <-ctx.Done():
-			err = ctx.Err()
-			return
-		}
-	}
-}
-
-// Create allocates new database.
-func Create(
-	meta types.ResourceMeta,
-	gasPrice uint64,
-	advancePayment uint64,
-	privateKey *asymmetric.PrivateKey,
-) (
-	dbID proto.DatabaseID, dsn string, err error,
-) {
-	var (
-		nonceReq   = new(types.NextAccountNonceReq)
-		nonceResp  = new(types.NextAccountNonceResp)
-		req        = new(types.AddTxReq)
-		resp       = new(types.AddTxResp)
-		clientAddr proto.AccountAddress
-	)
-	if clientAddr, err = crypto.PubKeyHash(privateKey.PubKey()); err != nil {
-		err = errors.Wrap(err, "get local account address failed")
-		return
-	}
-	// allocate nonce
-	nonceReq.Addr = clientAddr
-
-	if err = rpc.RequestBP(route.MCCNextAccountNonce.String(), nonceReq, nonceResp); err != nil {
-		err = errors.Wrap(err, "allocate create database transaction nonce failed")
-		return
-	}
-
-	req.Tx = types.NewCreateDatabase(&types.CreateDatabaseHeader{
-		Owner:          clientAddr,
-		ResourceMeta:   meta,
-		GasPrice:       gasPrice,
-		AdvancePayment: advancePayment,
-		TokenType:      types.Particle,
-		Nonce:          nonceResp.Nonce,
-	})
-
-	if err = req.Tx.Sign(privateKey); err != nil {
-		err = errors.Wrap(err, "sign request failed")
-		return
-	}
-
-	if err = rpc.RequestBP(route.MCCAddTx.String(), req, resp); err != nil {
-		err = errors.Wrap(err, "call create database transaction failed")
-		return
-	}
-
-	dbID = proto.FromAccountAndNonce(clientAddr, uint32(nonceResp.Nonce))
-	dsn = fmt.Sprintf("cql://%s", string(dbID))
-	return
 }
