+++ cmd/cql-fuse/block_test.go
@@ -47,8 +47,8 @@ import (
 	"testing"
 	"time"
 
-	bp "github.com/CovenantSQL/CovenantSQL/blockproducer"
 	"github.com/CovenantSQL/CovenantSQL/client"
+	"github.com/CovenantSQL/CovenantSQL/test"
 	"github.com/CovenantSQL/CovenantSQL/utils"
 	"github.com/CovenantSQL/CovenantSQL/utils/log"
 )
@@ -247,7 +247,7 @@ func initTestDB() (*sql.DB, func()) {
 	// wait for chain service
 	var ctx1, cancel1 = context.WithTimeout(context.Background(), 1*time.Minute)
 	defer cancel1()
-	err = bp.WaitBPChainService(ctx1, 3*time.Second)
+	err = test.WaitBPChainService(ctx1, 3*time.Second)
 	if err != nil {
 		log.Errorf("wait for chain service failed: %v", err)
 		return nil, stopNodes
